fine.

+++ TasteUp/src/main/java/com/group23/fooddelivery/persistence/dao/jdbc/PaninoDAOJDBC.java
@@ -70,7 +70,7 @@ public class PaninoDAOJDBC implements PaninoDAO{
 			ResultSet rs = st.executeQuery();
 			
 			if (rs.next()){
-				panino = new Panino(nome_panino, formato, rs.getInt("prezzo"), ingredienti, 
+				panino = new Panino(nome_panino, formato, rs.getFloat("prezzo"), ingredienti, 
 						rs.getString("path_image"));
 			}
